+++ CHANGELOG.md
@@ -11,5 +11,6 @@
 * Video
 
 ## [0.0.2] - 2021/05/27.
-* Button add CheckButton
-* add cascader
+* add CheckButton
+* add Cascader
+* add Text Marquee